AMP-61335 added node experiment codegen, codegen code cleanup
